TObjectPtr<T> used instead of TSubClassOf<T>!!!
When we try to access component with TSubClassOf<T>, it returns CDO (Class Default Object)! So we can not access what we assign in the Blueprint!

With TObjectPtr<T> we can access what we assign in the Blueprint. TObjectPtr<T> is not shown in the Blueprint so we take reference in BeginPlay!